+++ src/Table/shared/index.js
@@ -1,0 +1,14 @@
+export const TableHead = 'thead';
+
+export const TableBody = 'tbody';
+
+export * from './TableContext';
+export {default as TableWrapper} from './TableWrapper';
+export {default as TableRow} from './TableRow';
+export {default as TableColumnHeader} from './TableColumnHeader';
+export {default as TableRowHeader} from './TableRowHeader';
+export {default as TableCell} from './TableCell';
+export {default as TableBodyCell} from './TableBodyCell';
+export {default as TableCheckboxCell} from './TableCheckboxCell';
+export {default as RowSelectButton} from './RowSelectButton';
+export {default as InteractiveTableContent} from './InteractiveTableContent';